requirements.txt
alabaster==0.7.12
astroid==2.1.0
attrs==19.1.0
autopep8==1.4.3
Babel==2.7.0
beautifulsoup4==4.7.1
boto==2.49.0
boto3==1.9.184
botocore==1.12.184
breadability==0.1.20
Brotli==1.0.7
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cssselect==1.0.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.39.1
django-appconf==1.0.2
django-bootstrap-pagination==1.7.1
django-compat==1.0.15
django-crispy-forms==1.7.2
django-editor-widgets==2.0
django-favicon==0.1.3
django-favicon-plus==0.0.8
django-filebrowser-no-grappelli==3.7.8
django-filer==1.5.0
django-heroku==0.3.1
django-image-cropping==1.2.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.0.3
django-resized==0.3.9
django-secure==1.0.1
django-smartstaticfiles==0.3.2
django-social-share==1.3.2
django-sslserver==0.21
django-storages==1.7.1
django-template-utils==0.1.1
django-tinymce4-lite==1.7.5
django-widget-tweaks==1.4.5
docopt==0.6.2
docutils==0.14
easy-thumbnails==2.6
feedfinder2==0.0.4
feedparser==5.2.1
futures==3.1.1
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
jieba3k==0.35.1
Jinja2==2.10.1
jmespath==0.9.4
jsmin==2.2.2
lazy-object-proxy==1.3.1
lxml==4.3.4
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
newspaper3k==0.2.8
nltk==3.4.3
numpy==1.16.4
oauthlib==3.0.1
packaging==19.1
Pillow==6.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycountry==18.12.8
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1.1
redis==3.3.8
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.2.0
s3==3.0.0
s3transfer==0.2.1
six==1.12.0
snowballstemmer==1.9.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.2
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
sumy==0.8.1
tinysegmenter==0.3
tldextract==2.2.1
typed-ast==1.1.1
Unidecode==1.0.23
urllib3==1.25.3
Wand==0.5.7
whitenoise==4.1.3
wrapt==1.11.0
xmltodict==0.12.0
